split menu out as a component
