+++ src/App.js
@@ -1,20 +1,8 @@
-import React, { useState } from 'react';
-import MenuSections from './components/MenuSections';
-import ItemsList from './components/ItemsList';
-function App() {
-  const [menuSection, setMenuSection] = useState('skewers');
-  function updateMenuSection(id) {
-    setMenuSection(id);
-  }
+import Menu from './components/Menu';
+export default function App() {
   return (
-    <div className='App'>
-      <MenuSections
-        currentSection={menuSection}
-        updateMenuSection={updateMenuSection}
-      />
-      <ItemsList currentSection={menuSection} />
-    </div>
+    <>
+      <Menu />
+    </>
   );
 }
-
-export default App;
